refactor code to a error.c file and add a panic for emulator ref #226
